Register & Log in Function
I forgot to commit my first part of the project, what i did was
i created a function to register a user in offline mode by using the
switch case.

what i did after that was adding the log in function by using the same
switch case.

My next part of the project will be to add a error handling function to
see if a user already exists when creating a new user. I will also start
working on the trading part of the system.